Added authentification to API

+++ api/utils.py
@@ -1,13 +1,18 @@
 # pylint: disable=no-member
-from datetime import datetime
+from datetime import datetime, timedelta
 from os.path import join as path_join
 from typing import Callable, Dict, List, Optional, Type, Union
 
 import asyncpg
+import jwt
 from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security import OAuth2PasswordBearer
+from jwt import PyJWTError
 from passlib.context import CryptContext
 from pydantic import BaseModel
 from pytz import utc
+from starlette.requests import Request
+from starlette.status import HTTP_401_UNAUTHORIZED
 
 from bitcart import BTC
 
@@ -38,6 +43,54 @@ async def authenticate_user(username: str, password: str):
     return user
 
 
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/token")
+
+
+def create_access_token(
+    *, data: dict, expires_delta: timedelta = timedelta(minutes=15)
+):
+    to_encode = data.copy()
+    expire = now() + expires_delta
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(
+        to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM
+    )
+    return encoded_jwt
+
+
+class AuthDependency:
+    def __init__(self, enabled: bool = True):
+        self.enabled = enabled
+
+    async def __call__(self, request: Request):
+        if not self.enabled:
+            return None
+        token: str = await oauth2_scheme(request)
+        from . import schemes
+
+        credentials_exception = HTTPException(
+            status_code=HTTP_401_UNAUTHORIZED,
+            detail="Could not validate credentials",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+        try:
+            payload = jwt.decode(
+                token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
+            )
+            username: str = payload.get("sub")
+            if username is None:
+                raise credentials_exception
+            token_data = schemes.TokenData(username=username)
+        except PyJWTError:
+            raise credentials_exception
+        user = await models.User.query.where(
+            models.User.username == token_data.username
+        ).gino.first()
+        if user is None:
+            raise credentials_exception
+        return user
+
+
 HTTP_METHODS: List[str] = ["GET", "POST", "PUT", "PATCH", "DELETE"]
 
 
@@ -47,25 +100,31 @@ def model_view(
     orm_model,
     pydantic_model,
     create_model=None,
+    display_model=None,
     allowed_methods: List[str] = ["GET_ONE"] + HTTP_METHODS,
     custom_methods: Dict[str, Callable] = {},
     background_tasks_mapping: Dict[str, Callable] = {},
+    auth=True,
 ):
+    from . import schemes
+
+    display_model = pydantic_model if not display_model else display_model
+
     class PaginationResponse(BaseModel):
         count: int
         next: Optional[str]
         previous: Optional[str]
-        result: List[pydantic_model]  # type: ignore
+        result: List[display_model]  # type: ignore
 
     if not create_model:
         create_model = pydantic_model
     response_models: Dict[str, Type] = {
         "get": PaginationResponse,
-        "get_one": pydantic_model,
-        "post": pydantic_model,
-        "put": pydantic_model,
-        "patch": pydantic_model,
-        "delete": pydantic_model,
+        "get_one": display_model,
+        "post": display_model,
+        "put": display_model,
+        "patch": display_model,
+        "delete": display_model,
     }
 
     item_path = path_join(path, "{model_id}")
@@ -78,22 +137,43 @@ def model_view(
         "delete": item_path,
     }
 
-    async def get(pagination: pagination.Pagination = Depends()):
-        return await pagination.paginate(orm_model)
+    auth_dependency = AuthDependency(auth)
 
-    async def get_one(model_id: int):
-        item = await orm_model.get(model_id)
+    async def get(
+        pagination: pagination.Pagination = Depends(),
+        user: Union[None, schemes.User] = Depends(auth_dependency),
+    ):
+        if custom_methods.get("get"):
+            return await custom_methods["get"](pagination, user)
+        else:
+            return await pagination.paginate(orm_model)
+
+    async def get_one(
+        model_id: int, user: Union[None, schemes.User] = Depends(auth_dependency)
+    ):
+        if custom_methods.get("get_one"):
+            item = await custom_methods["get_one"](model_id, user)
+        else:
+            item = await orm_model.get(model_id)
         if not item:
             raise HTTPException(
                 status_code=404, detail=f"Object with id {model_id} does not exist!"
             )
         return item
 
+    create_dependency = (
+        auth_dependency if create_model != schemes.CreateUser else AuthDependency(False)
+    )
+
     async def post(
-        model: create_model  # type: ignore
+        model: create_model,  # type: ignore,
+        user: Union[None, schemes.User] = Depends(create_dependency),
     ):
         try:
-            obj = await orm_model.create(**model.dict())  # type: ignore
+            if custom_methods.get("post"):
+                obj = await custom_methods["post"](model, user)
+            else:
+                obj = await orm_model.create(**model.dict())  # type: ignore
         except (
             asyncpg.exceptions.UniqueViolationError,
             asyncpg.exceptions.NotNullViolationError,
@@ -104,10 +184,17 @@ def model_view(
             background_tasks_mapping["post"].send(obj.id)
         return obj
 
-    async def put(model_id: int, model: pydantic_model):  # type: ignore
+    async def put(
+        model_id: int,
+        model: pydantic_model,
+        user: Union[None, schemes.User] = Depends(auth_dependency),
+    ):  # type: ignore
         item = await get_one(model_id)
         try:
-            await item.update(**model.dict()).apply()  # type: ignore
+            if custom_methods.get("put"):
+                await custom_methods["put"](item, model, user)
+            else:
+                await item.update(**model.dict()).apply()  # type: ignore
         except (
             asyncpg.exceptions.UniqueViolationError,
             asyncpg.exceptions.NotNullViolationError,
@@ -116,12 +203,19 @@ def model_view(
             raise HTTPException(422, e.message)
         return item
 
-    async def patch(model_id: int, model: pydantic_model):  # type: ignore
+    async def patch(
+        model_id: int,
+        model: pydantic_model,
+        user: Union[None, schemes.User] = Depends(auth_dependency),
+    ):  # type: ignore
         item = await get_one(model_id)
         try:
-            await item.update(
-                **model.dict(skip_defaults=True)  # type: ignore
-            ).apply()
+            if custom_methods.get("patch"):
+                await custom_methods["patch"](item, model, user)
+            else:
+                await item.update(
+                    **model.dict(skip_defaults=True)  # type: ignore
+                ).apply()
         except (
             asyncpg.exceptions.UniqueViolationError,
             asyncpg.exceptions.NotNullViolationError,
@@ -130,16 +224,21 @@ def model_view(
             raise HTTPException(422, e.message)
         return item
 
-    async def delete(model_id: int):
+    async def delete(
+        model_id: int, user: Union[None, schemes.User] = Depends(auth_dependency)
+    ):
         item = await get_one(model_id)
-        await item.delete()
+        if custom_methods.get("delete"):
+            await custom_methods["delete"](item, user)
+        else:
+            await item.delete()
         return item
 
     for method in allowed_methods:
         method_name = method.lower()
         router.add_api_route(
             paths.get(method_name),  # type: ignore
-            custom_methods.get(method_name) or locals()[method_name],
+            locals()[method_name],
             methods=[method_name if method in HTTP_METHODS else "get"],
             response_model=response_models.get(method_name),
         )
